requirements.txt
contourpy==1.0.7
cycler==0.11.0
fonttools==4.39.4
greenlet==2.0.2
holidays==0.25
install==1.3.5
joblib==1.2.0
keras==2.12.0
kiwisolver==1.4.4
korean-lunar-calendar==0.3.1
matplotlib==3.7.1
MetaTrader5==5.0.45
mysql-connector-python==8.0.33
numpy==1.24.3
packaging==23.1
pandas==2.0.2
Pillow==9.5.0
protobuf==3.20.3
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
SQLAlchemy==2.0.16
threadpoolctl==3.1.0
typing_extensions==4.6.3
tzdata==2023.3
